Add serialization
Add the ability for a Lock object to be serialized down to a plain
UTF-8 string (JSON encoded). This lets it easily be passed between
systems.

Add the ability for a Session object to deserialize a serialized lock
and render it into a usable Lock. This requires that the technique and
backend used by the Session are the same as the ones that originally
were used to construct the lock.